requirements.txt
asgiref==3.7.2
Django==4.2.4
django-allauth==0.54.0
geographiclib==2.0
geopy==2.4.0
oauthlib==3.2.2
Pillow==10.0.0
PyJWT==2.8.0
python3-openid==3.2.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
